Crate servo_tracing

Structs

Fields 🔒

InstrumentConfiguration 🔒

Enums

Directive 🔒

Functions

instrument_internal 🔒

should_apply_instrument_method 🔒

Attribute Macros

instrument

Instruments a function with some sane defaults by automatically:

instrument_all

Instruments a whole impl block with the default servo tracing. See the documentation for #[servo_tracing::instrument]. If you want special treatment of a specific function you can override it with the normal tracing macro

skip

This is only useful for #[servo_tracing::instrument_all] and has no functionality except in that context.`