servo_tracing

Attribute Macro instrument_all

#[instrument_all]

Instruments a whole impl block with the default servo tracing. See the documentation for #[servo_tracing::instrument]. If you want special treatment of a specific function you can override it with the normal tracing macro

struct Foo {}
#[servo_tracing::instrument_all(skip_all))]
impl Foo {
    fn bar(&self) {}

    #[servo_tracing::instrument(name="Do baz")]
    fn baz(&self) {}

    fn bazz(&self) {}
}